refactor: overall refactoring
