+++ src/main/java/com/messenger/chatty/config/web/CorsMvcConfig.java
@@ -11,7 +11,6 @@ public class CorsMvcConfig implements WebMvcConfigurer {
     public void addCorsMappings(CorsRegistry corsRegistry) {
 
         corsRegistry.addMapping("/**")
-                .allowedOrigins("http://localhost:3000")
-                .exposedHeaders("Set-Cookie");
+                .allowedOrigins("http://localhost:3000");
     }
 }
