+++ src/main/java/com/messenger/chatty/controller/AuthController.java
@@ -2,7 +2,7 @@ package com.messenger.chatty.controller;
 
 
 import com.messenger.chatty.dto.request.LoginRequestDto;
-import com.messenger.chatty.dto.response.member.TokenResponseDto;
+import com.messenger.chatty.dto.response.auth.TokenResponseDto;
 import com.messenger.chatty.service.TokenService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -23,13 +23,13 @@ import org.springframework.web.bind.annotation.RestController;
 public class AuthController {
     private final TokenService tokenService;
 
-    @Operation(summary = "로그인",description = "리프레시 토큰은 쿠키에 발급되며 엑세스토큰은 Authorization 헤더에 발급됩니다.")
+    @Operation(summary = "로그인",description = "리프레시 토큰과 엑세스토큰이 body에 담겨 응답됩니다.")
     @PostMapping("/login")
     public ResponseEntity<Void> login(@RequestBody LoginRequestDto loginRequestDto)
     { // 스웨거 문서화를 위한 형식상 메서드
         return new ResponseEntity<>(HttpStatus.OK);
     }
-    @Operation(summary = "로그아웃",description = "헤더에 엑세스토큰이 담겨 있을 시에, 리프레시토큰이 disable 되며 정상 로그아웃됩니다.")
+    @Operation(summary = "로그아웃",description = "헤더에 엑세스토큰을 담아 요청을 보내면 리프레시토큰이 disable 되며 정상 로그아웃됩니다.")
     @PostMapping("/logout")
     public ResponseEntity<Void> logout(HttpServletRequest request, HttpServletResponse response)
     {
@@ -37,7 +37,7 @@ public class AuthController {
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    @Operation(summary = "엑세스토큰 및 리프레시 토큰 재발급",description = "기한이 짧은 엑세스 토큰이 만료 시 헤더에 리프레시토큰을 담아 요청을 보내세요.")
+    @Operation(summary = "엑세스토큰 및 리프레시 토큰 재발급",description = "기한이 짧은 엑세스 토큰이 만료 시 헤더에 리프레시토큰을 담아 요청을 보내세요. body에 토큰이 담겨 응답됩니다.")
     @PostMapping("/reissue")
     public TokenResponseDto reissue(HttpServletRequest request, HttpServletResponse response) {
             return  tokenService.reissueToken(request);
