+++ src/main/java/com/messenger/chatty/entity/RefreshToken.java
@@ -19,6 +19,7 @@ public class RefreshToken {
     // 다중 기기 로그인 등을 고려하면 나중에 멤버와의 일대다로 구성
     //@ManyToOne(fetch = FetchType.LAZY)
     // @JoinColumn(name = "member_id", nullable = false)
+    @Column(unique = true)
     private String username;
 
     @Column(nullable = false, unique = true)
