+++ src/main/java/com/messenger/chatty/repository/TokenRepository.java
@@ -4,6 +4,7 @@ import com.messenger.chatty.entity.RefreshToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TokenRepository  extends JpaRepository<RefreshToken,Long> {
-    Boolean existsByToken(String token);
-    void deleteByToken(String token);
+    Boolean existsByUsername(String username);
+    void deleteByUsername(String username);
+    RefreshToken findByUsername(String username);
 }
