+++ src/main/java/com/messenger/chatty/security/oauth2/Oauth2Response.java
@@ -6,4 +6,5 @@ public interface Oauth2Response {
     String getName();
     String getEmail();
     String getProfileImgURL();
+
 }
