+++ src/main/java/com/messenger/chatty/service/TokenServiceImpl.java
@@ -1,13 +1,11 @@
 package com.messenger.chatty.service;
-
-
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.exceptions.JWTVerificationException;
 import com.auth0.jwt.exceptions.TokenExpiredException;
 import com.auth0.jwt.interfaces.DecodedJWT;
-import com.messenger.chatty.dto.response.member.TokenResponseDto;
+import com.messenger.chatty.dto.response.auth.TokenResponseDto;
 import com.messenger.chatty.entity.RefreshToken;
 import com.messenger.chatty.repository.TokenRepository;
 import jakarta.servlet.http.HttpServletRequest;
@@ -47,14 +45,14 @@ public class TokenServiceImpl implements TokenService{
   @Override
   public TokenResponseDto reissueToken(HttpServletRequest request) {
       String token = getTokenFromRequest(request);
-      DecodedJWT decodedJWT = verifyAndDecodeToken(token, "refresh");
+      DecodedJWT decodedJWT = decodeToken(token, "refresh");
 
       String username = decodedJWT.getSubject();
       String role = decodedJWT.getClaim("role").asString();
-      TokenResponseDto tokenPair = generateTokenPair(username, role);
-      deleteRefreshToken(token);
-      saveRefreshToken(tokenPair.getRefresh_token(), username);
 
+      if(!tokenRepository.existsByUsername(username)) throw new IllegalStateException("갱신되기 이전 토큰입니다.");
+      TokenResponseDto tokenPair = generateTokenPair(username, role);
+      saveRefreshToken(tokenPair.getRefresh_token(), username);
       return tokenPair;
   }
 
@@ -62,39 +60,21 @@ public class TokenServiceImpl implements TokenService{
     @Override
     public void logout(HttpServletRequest request) {
         String token = getTokenFromRequest(request);
-        verifyAndDecodeToken(token, "refresh");
-        deleteRefreshToken(token);
+        DecodedJWT decodedJWT = decodeToken(token, "access");
+        String username = decodedJWT.getSubject();
+
+        // 일단 유저 한명당 디바이스에 관계없이 하나의 리프레시 토큰 만을 가질 수 있도록 함
+        tokenRepository.deleteByUsername(username);
     }
 
 
     @Override
     public void saveRefreshToken(String token,String username){
         RefreshToken refreshToken = RefreshToken.createTokenEntity(token, username);
+        // 존재한다면 삭제
+        tokenRepository.deleteByUsername(username);
         tokenRepository.save(refreshToken);
     }
-
-
-
-    public DecodedJWT verifyAndDecodeToken(String token, String tokenType){
-        DecodedJWT decodedJWT;
-        if (token == null) throw new IllegalStateException("토큰이 헤더에 없습니다.");
-        try {
-            decodedJWT = this.verifyJWT(token);
-        } catch (TokenExpiredException e) {
-            throw new IllegalStateException("토큰이 만료되었습니다.");
-        } catch (JWTVerificationException e) {
-            throw new IllegalStateException("토큰이 유효하지 않습니다.");
-        }
-
-        String category = decodedJWT.getClaim("category").asString();
-
-        if (!category.equals(tokenType)) throw new IllegalStateException("토큰 타입이 유효하지 않습니다.");
-        if(tokenType.equals("refresh") && !tokenRepository.existsByToken(token)) throw new IllegalStateException("갱신되기 이전 토큰입니다.");
-
-        return decodedJWT;
-
-    }
-
 
 
     public String getTokenFromRequest(HttpServletRequest request) {
@@ -104,6 +84,23 @@ public class TokenServiceImpl implements TokenService{
         }
         return authorization.split(" ")[1];
     }
+
+    public DecodedJWT decodeToken(String token, String tokenType){
+        DecodedJWT decodedJWT;
+        if (token == null) throw new IllegalStateException("토큰이 헤더에 없습니다.");
+        try {
+            decodedJWT = this.verifyJWT(token);
+        } catch (TokenExpiredException e) {
+            throw new IllegalStateException("토큰이 만료되었습니다.");
+        } catch (JWTVerificationException e) {
+            throw new IllegalStateException("토큰이 유효하지 않습니다.");
+        }
+        String category = decodedJWT.getClaim("category").asString();
+        if (!category.equals(tokenType)) throw new IllegalStateException("토큰 타입이 해당요청에 유효하지 않습니다.");
+        return decodedJWT;
+
+    }
+
 
     private DecodedJWT verifyJWT(String token)  {
         Algorithm algorithm = Algorithm.HMAC256(jwtKey.getBytes(StandardCharsets.UTF_8));
@@ -140,8 +137,5 @@ public class TokenServiceImpl implements TokenService{
     }
 
 
-    private void deleteRefreshToken(String token){
-        tokenRepository.deleteByToken(token);
-    }
 
 }
